requirements.txt
Flask>=2.0
Flask-Login
Flask-WTF
Flask-Migrate
Flask_SQLAlchemy
WTForms
email-validator
python-dotenv